Test/EXUI-4220 sync env bootstrap with Key Vault
- add a single repo-root .env.example for Playwright, API and local setup
- add Key Vault population scripts and keep the legacy Playwright wrapper
- add shared API env bootstrap so Node, API and Pact commands load the same .env
- document the env loading model and ignore .env_copy
