+++ .env.example
@@ -1,0 +1,210 @@
+# Repo-wide local env template for rpx-xui-webapp.
+# Copy to .env and populate tagged secrets from Azure Key Vault:
+#   yarn env:populate:aat
+#   yarn env:populate:demo
+#
+# If adding a new value, ensure Key Vault is updated and tagged with:
+#   e2e=ENV_VAR_NAME
+#
+# Notes:
+# - Keep placeholders only in this file. Do not commit live secret values.
+# - Blank values are expected to be filled by the Key Vault population flow or local overrides.
+
+# General
+TEST_ENV=aat
+TEST_URL=
+EXUI_BASE_URL=
+MANAGE_CASES_BASE_URL=
+BACKEND_BASE_URL=
+CITIZEN_FRONTEND_BASE_URL=
+JURISDICTION=PRIVATELAW
+CASE_TYPE=PRLAPPS
+
+# Node/API local defaults
+NODE_CONFIG_DIR=../config
+NODE_CONFIG_ENV=development
+ALLOW_CONFIG_MUTATIONS=1
+FEATURE_APP_INSIGHTS_ENABLED=false
+
+# IDAM
+IDAM_SECRET=
+IDAM_WEB_URL=
+IDAM_TESTING_SUPPORT_USERS_URL=
+IDAM_TESTING_SUPPORT_URL=
+IDAM_CITIZEN_USER_PASSWORD=
+IDAM_API_URL=
+IDAM_OAUTH2_SCOPE=openid profile roles
+IDAM_CLIENT_ID=
+CLIENT_ID=
+IDAM_RETURN_URL=
+SERVICES_IDAM_WEB_URL=
+SERVICES_IDAM_API_URL=
+SERVICES_IDAM_CLIENT_ID=xuiwebapp
+
+# CourtNav
+COURTNAV_SECRET=
+COURTNAV_USERNAME=
+COURTNAV_PASSWORD=
+COURTNAV_SUBSCRIPTION_KEY_CREATE_CASE=
+COURTNAV_CASE_URL=
+COURTNAV_DOC_URL=
+COURTNAV_SUBSCRIPTION_KEY_ADD_DOC=
+
+# S2S
+S2S_URL=
+S2S_SECRET=
+MICROSERVICE=xui_webapp
+S2S_MICROSERVICE_NAME=xui_webapp
+S2S_TOKEN=
+
+# CCD
+CCD_DATA_STORE_CLIENT_USERNAME=
+CCD_DATA_STORE_CLIENT_PASSWORD=
+CCD_DATA_STORE_CLIENT_ID=
+CCD_DATA_STORE_SECRET=
+CCD_DATA_STORE_URL=
+
+# Manage Case / downstream APIs
+MANAGE_CASE_REDIRECT_URI=
+PRL_COS_API_URL=
+MANAGE_ORG_API_PATH=
+RD_PROFESSIONAL_API_PATH=
+SERVICES_PRD_API=
+
+# Dynamic-user bootstrap
+CREATE_USER_SCOPE=openid profile roles create-user manage-user search-user
+CREATE_USER_CLIENT_ID=
+CREATE_USER_CLIENT_SECRET=
+CREATE_USER_BEARER_TOKEN=
+
+ORG_USER_ASSIGNMENT_BEARER_TOKEN=
+ORG_USER_ASSIGNMENT_USERNAME=
+ORG_USER_ASSIGNMENT_PASSWORD=
+IDAM_SOLICITOR_USER_PASSWORD=
+IDAM_CASEWORKER_DIVORCE_PASSWORD=
+ORG_USER_ASSIGNMENT_CLIENT_ID=
+ORG_USER_ASSIGNMENT_CLIENT_SECRET=
+ORG_USER_ASSIGNMENT_OAUTH2_SCOPE=openid profile roles
+ORG_USER_ASSIGNMENT_EXPECTED_EMAIL=
+ORG_USER_ASSIGNMENT_REDIRECT_URI=
+ORG_USER_ASSIGNMENT_UI_USER=
+ORG_USER_ASSIGNMENT_USER_ROLES=
+ORG_USER_ASSIGNMENT_MODE=auto
+ORG_USER_ASSIGNMENT_MODE_ORDER=internal,external
+ORG_USER_ASSIGNMENT_REQUEST_TIMEOUT_MS=20000
+ORG_ASSIGNMENT_ROLE_FILTER=true
+TEST_SOLICITOR_ORGANISATION_ID=
+
+# External professional-user assignment controls
+PROFESSIONAL_USER_ASSIGNMENT_MODE=external
+PROFESSIONAL_USER_ASSIGNMENT_FALLBACK_ORDER=external,internal
+PROFESSIONAL_USER_EXTERNAL_ASSIGNMENT_MAX_ATTEMPTS=3
+PROFESSIONAL_USER_EXTERNAL_ASSIGNMENT_RETRY_DELAY_MS=1000
+PROFESSIONAL_USER_EMAIL_DOMAIN=test.local
+PROFESSIONAL_USER_OUTPUT_CREATED_DATA=false
+PROFESSIONAL_USER_ASSIGNMENT_USE_MANAGE_ORG_PRIMARY=true
+PROFESSIONAL_USER_ENABLE_MANAGE_ORG_FALLBACK=true
+
+# Dynamic solicitor template and profile controls
+SOLICITOR_CASE_TYPE=PRLAPPS
+SOLICITOR_JURISDICTION=PRIVATELAW
+SOLICITOR_TEST_TYPE=provisioning
+SOLICITOR_ROLE_PROFILE=minimal
+PW_EMPLOYMENT_ASSIGNMENT_PREFLIGHT=true
+DYNAMIC_SOLICITOR_TEMPLATE_ROLES=
+DYNAMIC_PROD_LIKE_TEMPLATE_ROLES=
+
+# Case setup strategy for tests where case creation is not under UI test
+PW_E2E_CASE_SETUP_MODE=api-required
+PW_E2E_CASE_SETUP_API_ENDPOINT=
+PW_E2E_CASE_SETUP_TIMEOUT_MS=120000
+PW_E2E_CASE_SETUP_ALLOW_UI_FALLBACK=false
+
+# Guardrails for static/dynamic account usage in tests
+BLOCKED_STATIC_USERNAMES=
+ALLOW_CREDENTIAL_OUTPUT_IN_CI=false
+ALLOW_ORG_ASSIGNMENT_IDAM_OAUTH_IN_WORKER=true
+
+# Static aliases used by migrated E2E flows
+SOLICITOR_USERNAME=
+SOLICITOR_PASSWORD=
+PRL_SOLICITOR_USERNAME=
+PRL_SOLICITOR_PASSWORD=
+PRL_SOLICITOR2_USERNAME=
+PRL_SOLICITOR2_PASSWORD=
+PRL_SOLICITOR3_USERNAME=
+PRL_SOLICITOR3_PASSWORD=
+WA_SOLICITOR_USERNAME=
+WA_SOLICITOR_PASSWORD=
+NOC_SOLICITOR_USERNAME=
+NOC_SOLICITOR_PASSWORD=
+EVIDENCE_SOLICITOR_USERNAME=
+EVIDENCE_SOLICITOR_PASSWORD=
+EMPLOYMENT_DYNAMIC_SOLICITOR_USERNAME=
+EMPLOYMENT_DYNAMIC_SOLICITOR_PASSWORD=
+USER_WITH_FLAGS_USERNAME=
+USER_WITH_FLAGS_PASSWORD=
+SEARCH_EMPLOYMENT_CASE_USERNAME=
+SEARCH_EMPLOYMENT_CASE_PASSWORD=
+PROD_LIKE_USERNAME=
+PROD_LIKE_PASSWORD=
+CITIZEN_USERNAME=
+CITIZEN_PASSWORD=
+JUDGE_USERNAME=
+JUDGE_PASSWORD=
+CASEMANAGER_USERNAME=
+CASEMANAGER_PASSWORD=
+CASEWORKER_USERNAME=
+CASEWORKER_PASSWORD=
+COURT_ADMIN_STOKE_USERNAME=
+COURT_ADMIN_STOKE_PASSWORD=
+CIVIL_SOLICITOR_USERNAME=
+CIVIL_SOLICITOR_PASSWORD=
+
+# Playwright appTestConfig parity.
+# These keys map 1:1 to userIdentifier entries in playwright_tests_new/common/appTestConfig.ts
+# and are intended to be populated from Key Vault so local .env generation stays aligned with Jenkins.
+BOOKING_UI_FT_ON_USERNAME=
+BOOKING_UI_FT_ON_PASSWORD=
+CASEWORKER_GLOBALSEARCH_USERNAME=
+CASEWORKER_GLOBALSEARCH_PASSWORD=
+FPL_GLOBAL_SEARCH_USERNAME=
+FPL_GLOBAL_SEARCH_PASSWORD=
+HEARING_MANAGER_CR84_OFF_USERNAME=
+HEARING_MANAGER_CR84_OFF_PASSWORD=
+HEARING_MANAGER_CR84_ON_USERNAME=
+HEARING_MANAGER_CR84_ON_PASSWORD=
+IAC_ADMOFFICER_R1_USERNAME=
+IAC_ADMOFFICER_R1_PASSWORD=
+IAC_ADMOFFICER_R2_USERNAME=
+IAC_ADMOFFICER_R2_PASSWORD=
+IAC_CASEOFFICER_R1_USERNAME=
+IAC_CASEOFFICER_R1_PASSWORD=
+IAC_CASEOFFICER_R1_WITHPAGINATION_USERNAME=
+IAC_CASEOFFICER_R1_WITHPAGINATION_PASSWORD=
+IAC_CASEOFFICER_R1_WITHOUTPAGINATION_USERNAME=
+IAC_CASEOFFICER_R1_WITHOUTPAGINATION_PASSWORD=
+IAC_CASEOFFICER_R2_USERNAME=
+IAC_CASEOFFICER_R2_PASSWORD=
+IAC_JUDGE_WA_R1_USERNAME=
+IAC_JUDGE_WA_R1_PASSWORD=
+IAC_JUDGE_WA_R2_USERNAME=
+IAC_JUDGE_WA_R2_PASSWORD=
+IAC_JUDGE_WA_R2_CASEALLOCATOR_USERNAME=
+IAC_JUDGE_WA_R2_CASEALLOCATOR_PASSWORD=
+PROBATE_CW_USERNAME=
+PROBATE_CW_PASSWORD=
+RESTRICTED_CASE_ACCESS_OFF_USERNAME=
+RESTRICTED_CASE_ACCESS_OFF_PASSWORD=
+RESTRICTED_CASE_ACCESS_ON_USERNAME=
+RESTRICTED_CASE_ACCESS_ON_PASSWORD=
+RESTRICTED_CASE_FILE_VIEW_V1_1_OFF_USERNAME=
+RESTRICTED_CASE_FILE_VIEW_V1_1_OFF_PASSWORD=
+RESTRICTED_CASE_FILE_VIEW_V1_1_ON_USERNAME=
+RESTRICTED_CASE_FILE_VIEW_V1_1_ON_PASSWORD=
+STAFF_ADMIN_USERNAME=
+STAFF_ADMIN_PASSWORD=
+WA2_USERNAME=
+WA2_PASSWORD=
+WA2_GLOBAL_SEARCH_USERNAME=
+WA2_GLOBAL_SEARCH_PASSWORD=
