+++ api/package.json
@@ -11,11 +11,11 @@
     "ng": "ng",
     "build": "NODE_ENV=production webpack --config ./webpack-node.config.js",
     "build:dev": "NODE_ENV=development webpack --config ./webpack-node.config.js",
-    "coverage": "c8 --config ./.c8rc.json mocha --config ./test/.mocharc.json -r dotenv-extended/config",
+    "coverage": "c8 --config ./.c8rc.json mocha --config ./test/.mocharc.json -r ./setup-runtime-env.js -r dotenv-extended/config",
     "lint": "eslint .",
-    "watch": "nodemon -r dotenv-extended/config --watch '**/*.ts' --exec ts-node local.ts",
-    "test": "mocha --config ./test/.mocharc.json -r dotenv-extended/config",
+    "watch": "nodemon -r ./setup-runtime-env.js -r dotenv-extended/config --watch '**/*.ts' --exec ts-node local.ts",
+    "test": "mocha --config ./test/.mocharc.json -r ./setup-runtime-env.js -r dotenv-extended/config",
     "test:watch": "mocha --watch --recursive --watch-extensions ts -r ts-node/register",
-    "start": "node -r dotenv-extended/config ../dist/rpx-exui/api/server.bundle.js"
+    "start": "node -r ./setup-runtime-env.js -r dotenv-extended/config ../dist/rpx-exui/api/server.bundle.js"
   }
 }
