+++ package.json
@@ -35,19 +35,19 @@
     "nuke": "rm -rf dist",
     "pact-stub": "$(find . -name pact-stub-service | grep -e 'bin/pact-stub-service$' | head -n 1) ./api/test/pact/pacts/ --port 8080",
     "pre-push": "yarn lint && yarn build:prod && ng test --code-coverage",
-    "publish-pact": "NODE_PATH=. NODE_CONFIG_DIR=./config LOG_LEVEL=error dotenv_config_defaults=./api/.env.defaults NODE_OPTIONS=\"--disable-warning=MODULE_TYPELESS_PACKAGE_JSON --disable-warning=DEP0060\" mocha --timeout 15000 -r ts-node/register -r dotenv-extended/config api/test/pact/publish/publish.ts --reporter spec",
+    "publish-pact": "NODE_PATH=. LOG_LEVEL=error NODE_OPTIONS=\"--disable-warning=MODULE_TYPELESS_PACKAGE_JSON --disable-warning=DEP0060\" mocha --timeout 15000 -r ts-node/register -r api/test/setup-config-env.ts -r dotenv-extended/config api/test/pact/publish/publish.ts --reporter spec",
     "sonar-scan": "node_modules/sonar-scanner/bin/sonar-scanner",
     "start": "node ./dist/rpx-exui/api/server.bundle.js",
     "start:local": "cd api && yarn start",
     "start:node": "cd api && yarn watch",
-    "start:node:demo": "NODE_CONFIG_ENV=demo && cd api && yarn watch",
-    "start:node:dev": "NODE_CONFIG_ENV=development && cd api && yarn watch",
-    "start:node:development": "NODE_CONFIG_ENV=development && cd api && yarn watch",
-    "start:node:stub": "NODE_CONFIG_ENV=stub && cd api && yarn watch",
+    "start:node:demo": "cd api && NODE_CONFIG_ENV=demo yarn watch",
+    "start:node:dev": "cd api && NODE_CONFIG_ENV=development yarn watch",
+    "start:node:development": "cd api && NODE_CONFIG_ENV=development yarn watch",
+    "start:node:stub": "cd api && NODE_CONFIG_ENV=stub yarn watch",
     "test:node": "cd api && yarn coverage",
     "start:ng": "ng serve --port 3000 --host 0.0.0.0 --proxy-config proxy.config.json",
     "test": "cd api && yarn coverage",
-    "test:node:local": "cd api && NODE_CONFIG_DIR=../config NODE_CONFIG_ENV=development ALLOW_CONFIG_MUTATIONS=1 npx mocha --config ./test/.mocharc.json -r dotenv-extended/config",
+    "test:node:local": "cd api && ALLOW_CONFIG_MUTATIONS=1 npx mocha --config ./test/.mocharc.json -r ./setup-runtime-env.js -r dotenv-extended/config",
     "test:api": "PLAYWRIGHT_REPORT_FOLDER=functional-output/tests/playwright-api/odhin-report PLAYWRIGHT_REPORT_INDEX_FILENAME=xui-playwright-api.html npx playwright test --project node-api && node ./scripts/copy-odhin-report.js",
     "test:apiGateway": "yarn test:api",
     "test:coverage": "yarn test:coverage:ng",
@@ -64,6 +64,9 @@
     "test:fullfunctional": "yarn playwright install && npm-run-all -p -l -n -c test:api test:xuiIntegration",
     "test:playwrightE2E": "sh -c 'if [ \"${PLAYWRIGHT_SKIP_INSTALL}\" != \"true\" ]; then yarn playwright install; fi; npx playwright test --config=playwright.e2e.config.ts \"$@\"' --",
     "get-secrets": "node ./node_modules/@hmcts/playwright-common/dist/scripts/get-secrets.js",
+    "env:populate": "bash ./scripts/populate-env-from-keyvault.sh",
+    "env:populate:aat": "bash ./scripts/populate-env-from-keyvault.sh aat",
+    "env:populate:demo": "bash ./scripts/populate-env-from-keyvault.sh demo",
     "env:populate:playwright": "bash ./scripts/populate-playwright-env-from-keyvault.sh",
     "env:populate:playwright:aat": "bash ./scripts/populate-playwright-env-from-keyvault.sh aat",
     "env:populate:playwright:demo": "bash ./scripts/populate-playwright-env-from-keyvault.sh demo",
@@ -77,7 +80,7 @@
     "test:ng": "ng test",
     "test:ng:watch": "ng test --watch=true",
     "test:nodeMock": "nodemon test/nodeMock/app --standalone",
-    "test-pact": "NODE_PATH=. NODE_CONFIG_DIR=./config LOG_LEVEL=error dotenv_config_defaults=./api/.env.defaults NODE_OPTIONS=\"--disable-warning=MODULE_TYPELESS_PACKAGE_JSON --disable-warning=DEP0060\" mocha --timeout 10000 -r ts-node/register -r api/test/setup-config-env.ts -r dotenv-extended/config api/test/pact/**/*.spec.ts",
+    "test-pact": "NODE_PATH=. LOG_LEVEL=error NODE_OPTIONS=\"--disable-warning=MODULE_TYPELESS_PACKAGE_JSON --disable-warning=DEP0060\" mocha --timeout 10000 -r ts-node/register -r api/test/setup-config-env.ts -r dotenv-extended/config api/test/pact/**/*.spec.ts",
     "test:pact:run-and-publish": "yarn test-pact && yarn publish-pact",
     "test:setup:playwright-install-all": "npx playwright install --with-deps",
     "test:setup:playwright-install-chromium": "npx playwright install chromium",
