--- playwright_tests_new/.env.example
@@ -1,33 +1,0 @@
-# Playwright dynamic-user baseline env template.
-# Populate automatically from Key Vault:
-#   yarn env:populate:playwright:aat
-#   yarn env:populate:playwright:demo
-
-TEST_ENV=
-TEST_URL=
-
-IDAM_WEB_URL=
-SERVICES_IDAM_WEB_URL=
-IDAM_TESTING_SUPPORT_URL=
-IDAM_CLIENT_ID=
-IDAM_SECRET=
-S2S_URL=
-
-CREATE_USER_BEARER_TOKEN=
-ORG_USER_ASSIGNMENT_BEARER_TOKEN=
-
-ORG_USER_ASSIGNMENT_USERNAME=
-ORG_USER_ASSIGNMENT_PASSWORD=
-IDAM_SOLICITOR_USER_PASSWORD=
-IDAM_CASEWORKER_DIVORCE_PASSWORD=
-ORG_USER_ASSIGNMENT_CLIENT_ID=
-ORG_USER_ASSIGNMENT_CLIENT_SECRET=
-ORG_USER_ASSIGNMENT_OAUTH2_SCOPE=
-ORG_USER_ASSIGNMENT_EXPECTED_EMAIL=
-ORG_USER_ASSIGNMENT_REDIRECT_URI=
-ORG_USER_ASSIGNMENT_UI_USER=
-ORG_USER_ASSIGNMENT_USER_ROLES=
-
-TEST_SOLICITOR_ORGANISATION_ID=
-MANAGE_ORG_API_PATH=
-RD_PROFESSIONAL_API_PATH=
