+++ playwright_tests_new/README.md
@@ -100,8 +100,9 @@ yarn test:playwright:integration
 
 Use Key Vault tagged secrets to generate local `.env` for Playwright runs.
 
-- Template file: `playwright_tests_new/.env.example`
-- Script wrapper: `scripts/populate-playwright-env-from-keyvault.sh`
+- Template file: `.env.example`
+- Preferred script wrapper: `scripts/populate-env-from-keyvault.sh`
+- Compatibility wrapper: `scripts/populate-playwright-env-from-keyvault.sh`
 - Underlying helper: `@hmcts/playwright-common/dist/scripts/get-secrets.js`
 
 Tagging rule in Azure Key Vault:
@@ -112,16 +113,16 @@ Supported commands:
 
 ```bash
 # AAT
-yarn env:populate:playwright:aat
+yarn env:populate:aat
 
 # DEMO
-yarn env:populate:playwright:demo
+yarn env:populate:demo
 
 # Generic form (env + custom output file)
-yarn env:populate:playwright aat .env
+yarn env:populate aat .env
 
 # Direct helper form (vault, template, output)
-yarn get-secrets rpx-aat playwright_tests_new/.env.example .env
+yarn get-secrets rpx-aat .env.example .env
 ```
 
 Dynamic-user keys now available in Key Vault (`rpx-aat`, `rpx-demo`) and populated via tags:
@@ -144,6 +145,7 @@ Dynamic-user keys now available in Key Vault (`rpx-aat`, `rpx-demo`) and populat
 Notes:
 
 - Local dynamic-user creation requires F5 VPN (AAT/DEMO private services).
+- The repo-root `.env.example` is now the single source of truth for local `.env` generation.
 - Do not commit `.env`.
 
 ---
